pyproject.toml
[tool.poetry]
name = "python-learn"
version = "0.1.0"
description = ""
authors = ["zhangzhongguo <zhangzhongguo@kuaishou.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pygame = "^2.5.2"
requests = "^2.31.0"
pillow = "^10.1.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
